바탐시트 만들기

+++ app/src/main/java/com/rttgnu/readingteacherapp/LibraryModel.kt
@@ -5,11 +5,11 @@ import androidx.room.PrimaryKey
 
 @Entity(tableName = "library_table")
 data class LibraryModel (
-    @PrimaryKey val LibraryID: Int,
+    @PrimaryKey val LibraryID: Long,
     val LibraryName: String,
     val LibraryAuthor: String,
     val LibraryCover: Int,
-    val LibraryCategory: Int,
-    val LibrarySeriesNum: Int,
+    val LibraryCategory: Long,
+    val LibrarySeriesNum: Long,
     val LibraryDescription: String
 )
